Adds New Package Structure
