+++ lib/utilities/ApplicationConstants.dart
@@ -1,0 +1,3 @@
+library ApplicationConstants;
+
+const String APPLICATION_NAME = "ConcordiaGO";
