+++ lib/utilities/ConcordiaConstants.dart
@@ -1,0 +1,15 @@
+library ConcordiaConstants;
+
+// Loyola Buildings Names, Initials and Coordinates
+  //TODO
+
+// SGW Building Names, Initials and Coordinates
+  //TODO
+
+const String H_BUILDING_NAME = "Hall";
+const String H_BUILDING_INITIAL = "H";
+const double H_BUILDING_LATITUDE = 45.497113;
+const double H_BUILDING_LONGITUDE = -73.578791;
+
+const String FG_BUILDING_NAME = "Faubourg";
+const String FG_BUILDING_INITIAL = "FG";
